+++ example/lib/data/tu_chong_repository.dart
@@ -40,7 +40,7 @@ class TuChongRepository extends LoadingMoreBase<TuChongItem> {
   }
 
   @override
-  Future<bool> loadData([bool isloadMoreAction = false]) async {
+  Future<bool> loadData([bool isLoadMoreAction = false]) async {
     String url = '';
     if (isEmpty) {
       url = 'https://api.tuchong.com/feed-app';
@@ -84,4 +84,9 @@ class TuChongRepository extends LoadingMoreBase<TuChongItem> {
     }
     return isSuccess;
   }
+
+  // @override
+  // Iterable<TuChongItem> wrapData(Iterable<TuChongItem> source) {
+  //   return source.where((TuChongItem element) => element.imageCount == 1);
+  // }
 }