+++ example/lib/widget/item_builder.dart
@@ -128,7 +128,6 @@ Widget buildWaterfallFlowItem(BuildContext c, TuChongItem item, int index,
               ),
             );
             if (!konwSized) {
-              //todo: not work in web
               loadingWidget = AspectRatio(
                 aspectRatio: 1.0,
                 child: loadingWidget,